requirements.txt
funasr>=1.0.0
torch>=2.0.0
torchaudio
sounddevice
numpy
rumps
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
pynput
websockets
PyQt6
